Grab href as soon as extension is loaded
Since jupyterlab >=3, url parameters are consumed before the
session context is ready which breaks the functionality of
this extension. See issue:
https://github.com/jupyterlab/jupyterlab/issues/10275

As suggested by @mokkabonna, the parameters have not been consumed
immediately after loading the extension. At that time, jupyterlab
parameters are still present.

This change moves the reading of href to the suggested point and
adds filtering of built in query parameters.